src/bus/python/requirements.txt
nats-py>=2.8.0
